pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.black]
line-length = 100

[tool.poetry]
name = "defectdojo-importer"
version = "0.0.1"
description = "DefectDojo Importer Tool"
license = "Proprietary"
authors = [
    "Azunna Ikonne <ikonnea@gmail.com>"
]
maintainers = [
   "Azunna Ikonne <ikonnea@gmail.com>"
]

readme = "RELEASE.md"
repository = "https://github.com/1azunna/defectdojo-importer"

keywords = ["DefectDojo"]
include = ["CHANGELOG.md"]
classifiers = [
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]

packages = [
    { include = "defectdojo", from = "src" },
    { include = "integrations", from = "src" },
    { include = "models", from = "src" },
    { include = "common", from = "src" },
    { include = "importer", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4"
requests = "^2.32.5"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
responses = ">=0.23.1"
pytest-cov = "^4.1.0"
pytest-dotenv = "^0.5.2"
autohooks = "^23.4.0"
autohooks-plugin-black = "^23.4.0"
autohooks-plugin-pylint = "^23.4.0"
autohooks-plugin-pytest = "^23.4.0"

[tool.poetry.scripts]
defectdojo-importer = 'execute:main'

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
tag-branch = "main"
[tool.poetry-dynamic-versioning.substitution]
files = ["*/__init__.py"]

[tool.commitizen]
name = "cz_conventional_commits"
version_files = [
    "src/config/__init__.py",
    "pyproject.toml:version",
    "VERSION"
]
version_provider = "poetry"
annotated_tag = true
tag_format = "v$major.$minor.$patch$prerelease"
update_changelog_on_bump = true
pre_bump_hooks = [
  "scripts/remove-tags.sh"
]


[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.pylint", "autohooks.plugins.pytest"]
